chore: add middlewares on hours endpoints

+++ pkg/hour/service.go
@@ -1,6 +1,7 @@
 package hour
 
 import (
+	"fmt"
 	"github.com/gofiber/fiber/v2"
 	"github.com/gofiber/fiber/v2/middleware/session"
 	"log"
@@ -45,7 +46,19 @@ func (h Handler) GetHoursByShopId() fiber.Handler {
 
 func (h Handler) CreateHour() fiber.Handler {
 	return func(c *fiber.Ctx) error {
+		userRole := c.Locals("userRole")
+		userId := c.Locals("userId")
+
 		hour := parseHourBody(c)
+
+		shop, err := h.ShopStore.FindShopById(hour.ShopID)
+
+		if err != nil || (shop.OwnerID != userId && userRole != "admin") {
+			return c.Status(http.StatusUnauthorized).JSON(&fiber.Map{
+				"status":  "fail",
+				"message": "You are not authorized to create this resource",
+			})
+		}
 
 		createdHour, err := h.HourStore.CreateHour(*hour)
 		if err != nil {
@@ -77,10 +90,17 @@ func (h Handler) GetHourById() fiber.Handler {
 
 func (h Handler) DeleteHour() fiber.Handler {
 	return func(c *fiber.Ctx) error {
+		fmt.Println("DeleteHour")
+		isAllowedToUpdate := canUpdateHour(c, h)
+
+		if !isAllowedToUpdate {
+			return c.Status(http.StatusUnauthorized).JSON(&fiber.Map{
+				"status":  "fail",
+				"message": "You are not authorized to delete this resource",
+			})
+		}
+
 		id := c.Params("id")
-
-		println(id)
-
 		err := h.HourStore.DeleteHour(id)
 		if err != nil {
 			return c.Status(http.StatusInternalServerError).JSON(&fiber.Map{
@@ -95,6 +115,14 @@ func (h Handler) DeleteHour() fiber.Handler {
 
 func (h Handler) UpdateHour() fiber.Handler {
 	return func(c *fiber.Ctx) error {
+		isAllowedToUpdate := canUpdateHour(c, h)
+
+		if !isAllowedToUpdate {
+			return c.Status(http.StatusUnauthorized).JSON(&fiber.Map{
+				"status":  "fail",
+				"message": "You are not authorized to update this resource",
+			})
+		}
 		id := c.Params("id")
 		updatedHour := parseHourBody(c)
 
@@ -118,4 +146,27 @@ func parseHourBody(c *fiber.Ctx) *entities.Hour {
 	}
 
 	return hour
+}
+
+func canUpdateHour(c *fiber.Ctx, h Handler) bool {
+	userRole := c.Locals("userRole")
+	userId := c.Locals("userId")
+	id := c.Params("id")
+
+	if userRole == "admin" {
+		return true
+	}
+
+	hour, err := h.HourStore.GetHourById(id)
+	if err != nil {
+		return false
+	}
+
+	shop, err := h.ShopStore.FindShopById(hour.ShopID)
+
+	if err != nil || (shop.OwnerID != userId) {
+		return false
+	}
+
+	return true
 }
